ZIN-104
- Remove pointers utils
- Rework Event class
- Remove drawable class
- Add dragable "system" into gameplay loop

+++ Libs/WindowLib/include/Zinet/Window/ZtMousePositionEvent.hpp
@@ -8,6 +8,6 @@ namespace zt::wd
 {
 	struct ZINET_WINDOW_LAYER_API MousePositionEvent
 	{
-		Vector2f position;
+		Vector2d position;
 	};
 }
